+++ src/lib/components/modal/account/index.ts
@@ -1,3 +1,4 @@
+export * from "./account-qr-code";
 export * from "./EditSavedAccount";
 export * from "./RemoveSavedAccount";
 export * from "./SaveNewAccount";