+++ src/lib/pages/saved-accounts/components/SaveAccountButton.tsx
@@ -1,12 +1,16 @@
+import { Button } from "@chakra-ui/react";
 import { CustomIcon } from "lib/components/icon";
 import { SaveNewAccountModal } from "lib/components/modal";
 
 export const SaveAccountButton = () => (
   <SaveNewAccountModal
-    buttonProps={{
-      children: "Save new account",
-      leftIcon: <CustomIcon name="bookmark" />,
-      variant: "outline-primary",
-    }}
+    trigger={
+      <Button
+        as="button"
+        children="Save new account"
+        leftIcon={<CustomIcon name="bookmark" />}
+        variant="outline-primary"
+      />
+    }
   />
 );