+++ src/lib/components/move-abi/args-form/field/constants.ts
@@ -1,4 +1,11 @@
-export const UINT_TYPES = ["u8", "u16", "u32", "u64", "u128", "u256"];
+export const UINT_NUMBER_TYPES = ["u8", "u16", "u32"];
+
+export const UINT_STRING_TYPES = [
+  "u64",
+  "u128",
+  "u256",
+  "0x1::biguint::BigUint",
+];
 
 export const STRING_TYPE = "0x1::string::String";
 
@@ -9,7 +16,4 @@ export const FIXED_POINT_TYPES = [
   "0x1::fixed_point64::FixedPoint64",
 ];
 
-export const DECIMAL_TYPES = [
-  "0x1::decimal128::Decimal128",
-  "0x1::decimal256::Decimal256",
-];
+export const BIG_DECIMAL_TYPE = "0x1::bigdecimal::BigDecimal";