+++ src/lib/components/move-abi/args-form/field/utils/rules.ts
@@ -2,21 +2,24 @@ import type { Nullable, Option } from "lib/types";
 import type { FieldValues, UseControllerProps } from "react-hook-form";
 
 import big from "big.js";
-import { getArgType, getVectorElements } from "lib/utils";
+import { getArgType } from "lib/utils";
 import { parseInt } from "lodash";
 
 import {
-  DECIMAL_TYPES,
+  BIG_DECIMAL_TYPE,
   FIXED_POINT_TYPES,
   OBJECT_TYPE,
-  UINT_TYPES,
+  UINT_NUMBER_TYPES,
+  UINT_STRING_TYPES,
 } from "../constants";
 
-const validateNull = (v: Option<string>) =>
+const validateNull = (v: Nullable<unknown>) =>
   v !== null ? undefined : "cannot be null";
 
-const validateUint = (uintType: string) => (v: string) => {
+const validateUintNumber = (uintType: string) => (v: unknown) => {
   try {
+    if (typeof v !== "number") throw new Error();
+
     const value = big(v);
     const maxValue = big(2).pow(parseInt(uintType.slice(1)));
     if (value.lt(0) || value.gte(maxValue)) throw new Error();
@@ -25,21 +28,39 @@ const validateUint = (uintType: string) => (v: string) => {
     return `Input must be ‘${uintType}’`;
   }
 };
-const validateBool = (v: string) =>
-  v.toLowerCase() === "true" || v.toLowerCase() === "false"
-    ? undefined
-    : "Input must be ‘boolean’";
+
+const validateUintString = (uintType: string) => (v: unknown) => {
+  try {
+    if (typeof v !== "string") throw new Error();
+
+    const value = big(v);
+    const maxValue = big(2).pow(parseInt(uintType.slice(1)));
+    if (value.lt(0) || value.gte(maxValue)) throw new Error();
+    return undefined;
+  } catch {
+    return `Input must be ‘${uintType}’`;
+  }
+};
+
+const validateBool = (v: unknown) =>
+  typeof v === "boolean" ? undefined : "Input must be ‘boolean’";
 
 const validateAddress =
-  (isValidArgAddress: (input: string) => boolean) => (v: string) =>
-    isValidArgAddress(v) ? undefined : "Invalid address";
+  (isValidArgAddress: (input: string) => boolean) => (v: unknown) =>
+    typeof v === "string" && isValidArgAddress(v)
+      ? undefined
+      : "Invalid address";
 
 const validateObject =
-  (isValidArgObject: (input: string) => boolean) => (v: string) =>
-    isValidArgObject(v) ? undefined : "Invalid module address";
+  (isValidArgObject: (input: string) => boolean) => (v: unknown) =>
+    typeof v === "string" && isValidArgObject(v)
+      ? undefined
+      : "Invalid module address";
 
-const validateFixedPoint = (bcsFixedPointType: string) => (v: string) => {
+const validateFixedPoint = (bcsFixedPointType: string) => (v: unknown) => {
   try {
+    if (typeof v !== "string") throw new Error();
+
     const div = big(2).pow(
       parseInt(bcsFixedPointType.slice("fixed_point".length))
     );
@@ -52,56 +73,57 @@ const validateFixedPoint = (bcsFixedPointType: string) => (v: string) => {
   }
 };
 
-const validateDecimal = (bcsDecimalType: string) => (v: string) => {
-  const [, decimal] = v.split(".");
-  if (decimal && decimal.length > 18)
-    return "Decimal length must be less than 18";
+const validateBigDecimal = (v: unknown) => {
   try {
-    const value = big(v).times(big(10).pow(18));
-    const maxValue = big(2).pow(
-      parseInt(bcsDecimalType.slice("decimal".length))
-    );
-    if (value.lt(0) || value.gte(maxValue)) throw new Error();
+    if (typeof v !== "string") throw new Error();
+
+    const [, decimal] = v.split(".");
+    if (decimal && decimal.length > 18)
+      return "Decimal length must be less than 18";
+
     return undefined;
   } catch {
-    return `Input must be ‘${bcsDecimalType}’`;
+    return `Input must be ‘BigDecimal’`;
   }
 };
 
 const validateVector = (
-  v: string,
+  v: unknown,
   vectorType: string,
   isValidArgAddress: (input: string) => boolean,
   isValidArgObject: (input: string) => boolean
 ) => {
-  const value = v.trim();
-  if (!value.startsWith("[") || !value.endsWith("]"))
-    return "Input must be ‘vector’";
+  try {
+    if (!Array.isArray(v)) throw new Error();
+    const [, elementType] = vectorType.split(/<(.*)>/);
 
-  const [, elementType] = vectorType.split(/<(.*)>/);
+    // eslint-disable-next-line @typescript-eslint/no-unused-vars
+    let validateElement = (_v: unknown): Option<string> => undefined;
+    if (UINT_NUMBER_TYPES.includes(elementType))
+      validateElement = validateUintNumber(elementType);
+    else if (UINT_STRING_TYPES.includes(elementType))
+      validateElement = validateUintString(elementType);
+    else if (elementType === "bool") validateElement = validateBool;
+    else if (elementType === "address")
+      validateElement = validateAddress(isValidArgAddress);
+    else if (elementType.startsWith(OBJECT_TYPE))
+      validateElement = validateObject(isValidArgObject);
+    else if (FIXED_POINT_TYPES.includes(elementType))
+      validateElement = validateFixedPoint(getArgType(elementType));
+    else if (elementType === BIG_DECIMAL_TYPE)
+      validateElement = validateBigDecimal;
+    // TODO: handle Vector?
 
-  // eslint-disable-next-line @typescript-eslint/no-unused-vars
-  let validateElement = (_v: string): Option<string> => undefined;
-  if (UINT_TYPES.includes(elementType))
-    validateElement = validateUint(elementType);
-  if (elementType === "bool") validateElement = validateBool;
-  if (elementType === "address")
-    validateElement = validateAddress(isValidArgAddress);
-  if (elementType.startsWith(OBJECT_TYPE))
-    validateElement = validateAddress(isValidArgObject);
-  if (FIXED_POINT_TYPES.includes(elementType))
-    validateElement = validateFixedPoint(getArgType(elementType));
-  if (DECIMAL_TYPES.includes(elementType))
-    validateElement = validateDecimal(getArgType(elementType));
-  // TODO: handle Vector?
+    let error: Option<string>;
+    v.forEach((elementValue) => {
+      const err = validateElement(elementValue);
+      if (err !== undefined) error = `Invalid element: ${err}`;
+    });
 
-  let error: Option<string>;
-  getVectorElements(value).forEach((elementValue) => {
-    const res = validateElement(elementValue.trim());
-    if (res !== undefined) error = `Invalid element: ${res}`;
-  });
-
-  return error;
+    return error;
+  } catch {
+    return "Invalid vector format";
+  }
 };
 
 export const getRules = <T extends FieldValues>(
@@ -117,12 +139,20 @@ export const getRules = <T extends FieldValues>(
       null: validateNull,
     };
   }
-  if (UINT_TYPES.includes(type))
+  if (UINT_NUMBER_TYPES.includes(type))
     rules.validate = {
       ...rules.validate,
       [type]: (v: Nullable<string>) => {
         if (v === null) return undefined;
-        return validateUint(type)(v);
+        return validateUintNumber(type)(v);
+      },
+    };
+  if (UINT_STRING_TYPES.includes(type))
+    rules.validate = {
+      ...rules.validate,
+      [type]: (v: Nullable<string>) => {
+        if (v === null) return undefined;
+        return validateUintString(type)(v);
       },
     };
   if (type === "bool") {
@@ -162,13 +192,13 @@ export const getRules = <T extends FieldValues>(
       },
     };
   }
-  if (DECIMAL_TYPES.includes(type)) {
+  if (type === BIG_DECIMAL_TYPE) {
     const bcsType = getArgType(type);
     rules.validate = {
       ...rules.validate,
       [bcsType]: (v: Nullable<string>) => {
         if (v === null) return undefined;
-        return validateDecimal(bcsType)(v);
+        return validateBigDecimal(v);
       },
     };
   }