+++ src/lib/utils/abi.ts
@@ -5,8 +5,6 @@ import type {
   Nullable,
 } from "lib/types";
 
-import { BCS } from "@alleslabs/legacy-bcs";
-
 export const checkAvailability = (fn: ExposedFunction) =>
   fn.is_view || fn.is_entry;
 
@@ -72,70 +70,35 @@ export const getAbiInitialData = (length: number) =>
 // ------------------------------------------//
 // -----------------MOVE ARGS----------------//
 // ------------------------------------------//
-export const getVectorElements = (value: string) => {
-  const trimmed = value.split(/\[(.*)\]/)[1].trim();
-  if (trimmed.length === 0) return [];
-  return trimmed.split(",").map((element) => element.trim());
-};
 
 export const getArgType = (argType: string) =>
   argType
-    .replace("0x1::string::String", BCS.STRING)
-    .replace("0x1::option::Option", BCS.OPTION)
-    .replace("0x1::object::Object", BCS.OBJECT)
-    .replace("0x1::fixed_point32::FixedPoint32", BCS.FIXED_POINT32)
-    .replace("0x1::fixed_point64::FixedPoint64", BCS.FIXED_POINT64)
-    .replace("0x1::decimal128::Decimal128", BCS.DECIMAL128)
-    .replace("0x1::decimal256::Decimal256", BCS.DECIMAL256);
+    .replace("0x1::string::String", "string")
+    .replace("0x1::option::Option", "option")
+    .replace("0x1::object::Object", "object")
+    .replace("0x1::biguint::BigUint", "biguint")
+    .replace("0x1::fixed_point32::FixedPoint32", "fixed_point32")
+    .replace("0x1::fixed_point64::FixedPoint64", "fixed_point64")
+    .replace("0x1::bigdecimal::BigDecimal", "bigdecimal");
 
-const getChildType = (parentType: string, type: string) => {
-  const [pType, childType] = type.split(/<(.*)>/);
-  return pType === parentType ? childType : undefined;
-};
-const getArgValue = ({
-  type,
-  value,
-}: {
-  type: string;
-  value: Nullable<string>;
-}) => {
-  try {
-    if (value === null) return null;
-    if (type.startsWith("vector")) {
-      const elementType = getChildType("vector", type);
-      const elements = getVectorElements(value);
-      return elementType === "bool"
-        ? elements.map((element) => element.toLowerCase() === "true")
-        : elements;
-    }
-    if (type === "bool" || getChildType("0x1::option::Option", type) === "bool")
-      return value.toLowerCase() === "true";
-    return value.trim();
-  } catch {
-    return "";
-  }
+const serializeArgJson = (arg: { type: string; value: Nullable<unknown> }) => {
+  return JSON.stringify(arg.value);
 };
 
-const bcs = BCS.getInstance();
-
-const serializeArg = (arg: { type: string; value: Nullable<string> }) => {
-  const bufferSize = Math.max(1024, 2 * (arg.value ?? "").length);
-  try {
-    const argType = getArgType(arg.type);
-    const argValue = getArgValue(arg);
-    return bcs.serialize(argType, argValue, bufferSize);
-  } catch {
-    return "";
-  }
-};
-
-export const serializeAbiData = (fn: ExposedFunction, abiData: AbiFormData) => {
+export const serializeAbiDataJson = (
+  fn: ExposedFunction,
+  abiData: AbiFormData
+) => {
   const serializedArgs = fn.params.map((type, index) =>
-    serializeArg({ type, value: abiData.args[index] })
+    serializeArgJson({ type, value: abiData.args[index] })
+  );
+
+  const serializedTypeArgs = fn.generic_type_params.map(
+    (_, index) => abiData.typeArgs[index] ?? ""
   );
 
   return {
     args: serializedArgs,
-    typeArgs: fn.generic_type_params.map((_, index) => abiData.typeArgs[index]),
+    typeArgs: serializedTypeArgs,
   };
 };