+++ src/lib/utils/formatter/token.ts
@@ -105,6 +105,8 @@ export const formatUTokenWithPrecision = (
  */
 export const formatPrice = (value: USD<BigSource>): string => {
   try {
+    if (big(value).eq(0)) return "$ - ";
+
     const price = big(value);
     const lowestThreshold = 0.000001;