+++ CHANGELOG.md
@@ -41,6 +41,8 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ### Improvements
 
+- [#1346](https://github.com/alleslabs/celatone-frontend/pull/1346) Show $- for zero value instead
+
 ### Bug fixes
 
 - [#1343](https://github.com/alleslabs/celatone-frontend/pull/1343) Fix upgrade policy issue when publishing a module