+++ CHANGELOG.md
@@ -53,6 +53,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 - [#1347](https://github.com/alleslabs/celatone-frontend/pull/1347) Remove unused is move upgraded flag
 - [#1346](https://github.com/alleslabs/celatone-frontend/pull/1346) Show $- for zero value instead
+- [#1332](https://github.com/alleslabs/celatone-frontend/pull/1332) Add Abi encode with JSON serialize
 
 ### Bug fixes
 
@@ -83,7 +84,6 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ### Improvements
 
-- [#1332](https://github.com/alleslabs/celatone-frontend/pull/1332) Add Abi encode with JSON serialize
 - [#1328](https://github.com/alleslabs/celatone-frontend/pull/1328) Remove hot fix increment last byte for blocks
 - [#1318](https://github.com/alleslabs/celatone-frontend/pull/1318) Update ESLint rules
 - [#1327](https://github.com/alleslabs/celatone-frontend/pull/1327) Support show scrollbar when table overflow